+++ crates/enclave-server/src/key_manager/key_manager.rs
@@ -28,7 +28,8 @@ impl KeyPurpose {
         match self {
             KeyPurpose::Snapshot => "snapshot",
             KeyPurpose::RngPrecompile => "rng-precompile",
-            _ => panic!("unimplimented key purpose"), // TODO: should not panic
+            KeyPurpose::Secp256k1 => "secp256k1",
+            KeyPurpose::Schnorrkel => "schnorrkel",
         }
     }
 
@@ -127,30 +128,6 @@ mod tests {
     use super::*;
     use serial_test::serial;
 
-    #[test]
-    fn test_secret_from_vec_valid() {
-        let vec_32 = vec![1u8; 32];
-        let secret = Secret::from_vec(vec_32).unwrap();
-        assert_eq!(secret.as_ref().len(), 32);
-    }
-
-    #[test]
-    fn test_secret_from_vec_invalid_length() {
-        let vec_16 = vec![1u8; 16];
-        let res = Secret::from_vec(vec_16);
-
-        assert!(res.is_err(), "Expected error for invalid secret size");
-
-        if let Err(e) = res {
-            let msg = e.to_string();
-            assert!(
-                msg.contains("Invalid secret size"),
-                "Unexpected error message: {}",
-                msg
-            );
-        }
-    }
-
     // #[test]
     // fn test_key_manager_direct_constructor() {
     //     let master_key_bytes = [0u8; 32];
@@ -167,5 +144,16 @@ mod tests {
         let key_a = key_manager.get_key(KeyPurpose::Snapshot).unwrap();
         let key_b = key_manager.get_key(KeyPurpose::Snapshot).unwrap();
         assert_eq!(key_a.bytes, key_b.bytes);
+    }
+
+    #[test]
+    fn test_all_purpose_keys_are_initialized() {
+        let master_key_bytes = [0u8; 32];
+        let key_manager = KeyManager::new(master_key_bytes).unwrap();
+        
+        for purpose in KeyPurpose::iter() {
+            let key = key_manager.get_key(purpose).unwrap();
+            assert!(!key.bytes.is_empty());
+        }
     }
 }
